# Optimize encoding of composites